修改IF_IDF类型，完成step8

+++ src/main/java/org/example/Document.java
@@ -9,7 +9,7 @@ public class Document {
 
     String text;
     Map<String, Integer> termMap = new HashMap<>();
-    Map<String, Integer> IF_IDFMap = new HashMap<>();
+    Map<String, Double> IF_IDFMap = new HashMap<>();
 
     /**
      * 读取文件并初始化文件类
